// Source/Portfolio.Web/ClientApp/src/api/models/contentDto.ts
/**
 * Generated by orval v7.19.0 🍺
 * Do not edit manually.
 * Portfolio.Api
 * OpenAPI spec version: 1.0
 */
import type { HeroDto } from './heroDto';
import type { SkillCategoryDto } from './skillCategoryDto';
import type { ProjectDto } from './projectDto';
import type { ExperienceDto } from './experienceDto';
import type { EducationDto } from './educationDto';
import type { FactDto } from './factDto';
import type { ContactDto } from './contactDto';

export interface ContentDto {
  hero?: HeroDto;
  /** @nullable */
  skillsCategories?: SkillCategoryDto[] | null;
  /** @nullable */
  projects?: ProjectDto[] | null;
  /** @nullable */
  experiences?: ExperienceDto[] | null;
  /** @nullable */
  educations?: EducationDto[] | null;
  /** @nullable */
  facts?: FactDto[] | null;
  /** @nullable */
  contact?: ContactDto[] | null;
}